+++ worker/src/handlers/builds/build.ts
@@ -6,9 +6,9 @@ import { success } from '~/api/api';
 import * as errors from '~/api/errors';
 import Constants from '~/shared/utils/constants';
 import { Pages } from '~/shared/utils/routes';
-import BuildStore from '~/store/builds';
-import ProjectStore from '~/store/projects';
-import ReleaseChannelStore from '~/store/releaseChannels';
+import BuildStore from '~/store/BuildStore';
+import ProjectStore from '~/store/ProjectStore';
+import ReleaseChannelStore from '~/store/ReleaseChannelStore';
 import { Ctx } from '~/types/hono';
 import { getBuildId, getFilePath } from '~/utils/build';
 import { sha256 } from '~/utils/crypto';
